requirements.txt
asgiref==3.4.1
Django==3.2.9
django-crispy-forms==1.13.0
numpy==1.21.4
opencv-python==4.5.3.56
pafy==0.5.5
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
youtube-dl==2021.6.6
